updating the candidates array in the active election
