+++ client/store/candidate.js
@@ -2,7 +2,6 @@ import axios from 'axios';
 
 //Action Types
 const GET_CANDIDATES = 'GET_CANDIDATES';
-const POST_NEW_CANDIDATE = 'POST_NEW_CANDIDATE';
 
 //Action Creators
 const getCandidates = (candidates) => {
@@ -29,23 +28,4 @@ export const fetchCandidates = (election) => {
   }
 };
 
-export const postNewCandidate = (newCandidateObj, electionId) => {
-  return dispatch => {
-    axios.post(`/api/candidates/${electionId}`, newCandidateObj)
-      .then(res => res.data)
-      .then(created => console.log("new candidate posted! ", created))
-      .catch(console.error);
-  }
-};
 //do we need an action creator to update the candidates array with the new candidate now?
-
-//Reducer
-export function candidatesReducer(candidates = {}, action) {
-  switch (action.type) {
-    case GET_CANDIDATES:
-      return action.candidates
-    default:
-      return candidates
-  }
-}
-
