+++ client/store/index.js
@@ -4,10 +4,8 @@ import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
 import user from './user'
 import { activeElectionReducer, electionsReducer, blockchainElectionReducer } from './election';
-// import { giveWatchPartyCountsReducer } from './watch-party';
-import { candidatesReducer } from './candidate';
 
-const reducer = combineReducers({ user, activeElection: activeElectionReducer, elections: electionsReducer, blockchainElections: blockchainElectionReducer, candidates: candidatesReducer })
+const reducer = combineReducers({ user, activeElection: activeElectionReducer, elections: electionsReducer, blockchainElections: blockchainElectionReducer })
 const middleware = composeWithDevTools(applyMiddleware(
   thunkMiddleware,
   createLogger({collapsed: true})
